+++ database/migrations/2025_08_19_055831_add_status_to_vendor_register_lists_table.php
@@ -11,10 +11,12 @@ return new class extends Migration
      */
     public function up(): void
     {
-        Schema::table('vendor_register_lists', function (Blueprint $table) {
+        if (!Schema::hasColumn('vendor_register_lists', 'status')) {
+            Schema::table('vendor_register_lists', function (Blueprint $table) {
             $table->string('status')->default('pending')->after('business_id');
               // status: pending, approve, reject
         });
+        }
     }
 
     /**
@@ -22,8 +24,10 @@ return new class extends Migration
      */
     public function down(): void
     {
-        Schema::table('vendor_register_lists', function (Blueprint $table) {
+        if (!Schema::hasColumn('vendor_register_lists', 'status')) {
+            Schema::table('vendor_register_lists', function (Blueprint $table) {
             $table->dropColumn('status');
         });
+        }
     }
 };
